Move ialarm coordinator to separate module
